Create visual tests in the Timeline (#3593)
* Create visual tests in the Timeline

In this PR we are creating visual tests in the TimelineUi component.
After creating several mock classes, we were able to test comparisons
between number of major ticks, number of minor ticks and labels for
different ranges. We were also able to test number of characters in the
labels, and positions of objects.

As a result, one bug was found and fixed (the margin below timeline was
outside of the kZTimeline ranges and on top of the incomplete data
intervals: http://screenshot/56bT7Cr7b5aE5GM.

Also, we found something we previously now (small screens works a bit
weird). This is a small issue that need a bit of a refactor to be fixed
and not a big priority but it was good that it was catched in
these unit tests. The issue is that screens below 700 pixels width
sometimes has intersecting labels (actually it's more like 500 but the
MockTextRenderer is estimating every character as a 'w').

Finally, we don't expect in really small screens to force to have at
least 2 labels.

Bug: http://b/218302149.

* Fixup: Gcc presubmit

* Fixup: Addressing comments

* Fixup: Erase annonymous namespace

* Fixup: comments

+++ src/OrbitGl/MockTimelineInfo.h
@@ -25,7 +25,7 @@ class MockTimelineInfo : public TimelineInfoInterface {
     return GetWorldFromUs(GetUsFromTick(time));
   }
   [[nodiscard]] float GetWorldFromUs(double micros) const override {
-    return (micros - GetMinTimeUs()) * width_ / GetTimeWindowUs();
+    return static_cast<float>((micros - GetMinTimeUs()) * width_ / GetTimeWindowUs());
   }
   [[nodiscard]] uint64_t GetTickFromWorld(float world_x) const override {
     double ratio = world_x / width_;
